requirements.txt
# OpenAgentFlow 依赖
# 核心依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0

# AI相关
openai>=1.0.0
langchain>=0.0.340
langchain-openai>=0.0.2

# 数据库
alembic>=1.12.0
sqlite3  # 系统自带

# 工具集成 (可选)
# 飞书: 通过API调用
# 钉钉: 通过API调用
# 企业微信: 通过API调用

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# 其他
python-multipart>=0.0.6
email-validator>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0